refactor(Worklets): move ui scheduling
