+++ packages/react-native-reanimated/src/runtimes.ts
@@ -4,12 +4,12 @@ import type { WorkletFunction } from './commonTypes';
 import { ReanimatedError, registerReanimatedError } from './errors';
 import { setupCallGuard, setupConsole } from './initializers';
 import { registerLoggerConfig } from './logger';
-import { ReanimatedModule } from './ReanimatedModule';
 import { shouldBeUseWeb } from './PlatformChecker';
 import {
   makeShareableCloneOnUIRecursive,
   makeShareableCloneRecursive,
 } from './shareables';
+import { WorkletsModule } from './worklets';
 
 const SHOULD_BE_USE_WEB = shouldBeUseWeb();
 
@@ -43,7 +43,7 @@ export function createWorkletRuntime(
   // Assign to a different variable as __reanimatedLoggerConfig is not a captured
   // identifier in the Worklet runtime.
   const config = __reanimatedLoggerConfig;
-  return ReanimatedModule.createWorkletRuntime(
+  return WorkletsModule.createWorkletRuntime(
     name,
     makeShareableCloneRecursive(() => {
       'worklet';
@@ -86,7 +86,7 @@ export function runOnRuntime<Args extends unknown[], ReturnValue>(
       );
   }
   return (...args) =>
-    ReanimatedModule.scheduleOnRuntime(
+    WorkletsModule.scheduleOnRuntime(
       workletRuntime,
       makeShareableCloneRecursive(() => {
         'worklet';
