+++ packages/react-native-reanimated/src/worklets/WorkletsModule/NativeWorklets.ts
@@ -2,8 +2,9 @@
 import { getValueUnpackerCode } from '../valueUnpacker';
 import { WorkletsTurboModule } from '../../specs';
 import { ReanimatedError } from '../../errors';
-import type { IWorkletsModule } from '../../commonTypes';
+import type { IWorkletsModule, ShareableRef } from '../../commonTypes';
 import type { WorkletsModuleProxy } from './workletsModuleProxy';
+import type { WorkletRuntime } from '../../runtimes';
 
 export function createNativeWorkletsModule(): IWorkletsModule {
   return new NativeWorklets();
@@ -26,8 +27,8 @@ See https://docs.swmansion.com/react-native-reanimated/docs/guides/troubleshooti
     this.#workletsModuleProxy = global.__workletsModuleProxy;
   }
 
-  makeShareableClone<T>(
-    value: T,
+  makeShareableClone<TValue>(
+    value: TValue,
     shouldPersistRemote: boolean,
     nativeStateSource?: object
   ) {
@@ -35,6 +36,30 @@ See https://docs.swmansion.com/react-native-reanimated/docs/guides/troubleshooti
       value,
       shouldPersistRemote,
       nativeStateSource
+    );
+  }
+
+  scheduleOnUI<TValue>(shareable: ShareableRef<TValue>) {
+    return this.#workletsModuleProxy.scheduleOnUI(shareable);
+  }
+
+  executeOnUIRuntimeSync<TValue, TReturn>(
+    shareable: ShareableRef<TValue>
+  ): TReturn {
+    return this.#workletsModuleProxy.executeOnUIRuntimeSync(shareable);
+  }
+
+  createWorkletRuntime(name: string, initializer: ShareableRef<() => void>) {
+    return this.#workletsModuleProxy.createWorkletRuntime(name, initializer);
+  }
+
+  scheduleOnRuntime<T>(
+    workletRuntime: WorkletRuntime,
+    shareableWorklet: ShareableRef<T>
+  ) {
+    return this.#workletsModuleProxy.scheduleOnRuntime(
+      workletRuntime,
+      shareableWorklet
     );
   }
 }
