+++ packages/react-native-reanimated/src/worklets/WorkletsModule/workletsModuleProxy.ts
@@ -1,6 +1,7 @@
 'use strict';
 
 import type { ShareableRef } from '../../commonTypes';
+import type { WorkletRuntime } from '../../runtimes';
 
 /** Type of `__workletsModuleProxy` injected with JSI. */
 export interface WorkletsModuleProxy {
@@ -9,4 +10,20 @@ export interface WorkletsModuleProxy {
     shouldPersistRemote: boolean,
     nativeStateSource?: object
   ): ShareableRef<TValue>;
+
+  scheduleOnUI<TValue>(shareable: ShareableRef<TValue>): void;
+
+  executeOnUIRuntimeSync<TValue, TReturn>(
+    shareable: ShareableRef<TValue>
+  ): TReturn;
+
+  createWorkletRuntime(
+    name: string,
+    initializer: ShareableRef<() => void>
+  ): WorkletRuntime;
+
+  scheduleOnRuntime<TValue>(
+    workletRuntime: WorkletRuntime,
+    worklet: ShareableRef<TValue>
+  ): void;
 }
